+++ test.js
@@ -2,6 +2,11 @@ var clog = console.log;
 var nomo = require('./index.js').start();
 nomo.profiler.start();
 
+
+//
+// -- Test console functions --
+//
+
 function logObject() {
   clog('Sending object to clients');
   console.log({key1: 'hi', key2: ['deep', 'object']});
@@ -20,6 +25,8 @@ console.log({
     { nested: function() {} }
   ]
 });
+
+console.dir({test: 'hi', test1: {nested: 'object'}, test2: function() { alert('hi'); }});
 
 
 //